+++ neoforge/src/main/java/de/cech12/usefulhats/UsefulHatsNeoForgeMod.java
@@ -1,5 +1,6 @@
 package de.cech12.usefulhats;
 
+import de.cech12.usefulhats.compat.AccessoriesCompat;
 import de.cech12.usefulhats.compat.Baubles2Compat;
 import de.cech12.usefulhats.init.ModCreativeTabs;
 import de.cech12.usefulhats.init.ModItems;
@@ -26,12 +27,10 @@ public class UsefulHatsNeoForgeMod {
 
     @SubscribeEvent
     public static void registerCapabilities(RegisterCapabilitiesEvent event) {
-        /* TODO
-        if (Services.PLATFORM.isModLoaded(AccessoriesCompat.MOD_ID)) {
+        if (Services.PLATFORM.isModLoaded(Constants.ACCESSORIES_MOD_ID)) {
             AccessoriesCompat.register();
         }
-         */
-        if (Services.PLATFORM.isModLoaded(Baubles2Compat.MOD_ID)) {
+        if (Services.PLATFORM.isModLoaded(Constants.BAUBLES_2_MOD_ID)) {
             Baubles2Compat.register(event);
         }
     }