+++ neoforge/src/main/java/de/cech12/usefulhats/client/UsefulHatsNeoForgeClientEvents.java
@@ -1,8 +1,8 @@
 package de.cech12.usefulhats.client;
 
 import de.cech12.usefulhats.Constants;
+import de.cech12.usefulhats.client.compat.AccessoriesClientCompat;
 import de.cech12.usefulhats.client.compat.Baubles2ClientCompat;
-import de.cech12.usefulhats.compat.Baubles2Compat;
 import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.Services;
 import net.minecraft.client.Minecraft;
@@ -45,12 +45,10 @@ public class UsefulHatsNeoForgeClientEvents {
 
     @SubscribeEvent
     public static void onClientInit(final FMLClientSetupEvent event) {
-        /* TODO
-        if (Services.PLATFORM.isModLoaded(AccessoriesCompat.MOD_ID)) {
+        if (Services.PLATFORM.isModLoaded(Constants.ACCESSORIES_MOD_ID)) {
             AccessoriesClientCompat.register();
         }
-         */
-        if (Services.PLATFORM.isModLoaded(Baubles2Compat.MOD_ID)) {
+        if (Services.PLATFORM.isModLoaded(Constants.BAUBLES_2_MOD_ID)) {
             Baubles2ClientCompat.register();
         }
     }